기존 WineDetail 컴포넌트에서 fetch를 함 -> 별도의 useFetchWine 훅으로 분리하였으며 store와 연동
